# docs/protocol/http/openapi-v1.json
{
  "components": {
    "parameters": {
...
      "RequestId": {
        "in": "header",
        "name": "X-Request-ID",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "responses": {
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "Stable typed failure.",
        "headers": {
          "X-Request-ID": {
            "description": "Matches the request and JSON body request_id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ConnectionStatus": {
        "additionalProperties": false,
        "description": "对桌面端公开的 SSH 连接状态快照。",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "host_key_candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HostKeyCandidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "recoverable": {
            "title": "Recoverable",
            "type": "boolean"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "state": {
            "enum": [
              "DISCONNECTED",
              "CONNECTING",
              "HOST_KEY_REQUIRED",
              "READY",
              "CLOSING",
              "FAILED"
            ],
            "title": "State",
            "type": "string"
          },
          "trusted_fingerprint_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trusted Fingerprint Sha256"
          }
        },
        "required": [
          "connection_id",
          "state",
          "session_id",
          "error_code",
          "recoverable",
          "correlation_id",
          "host_key_candidate"
        ],
        "title": "ConnectionStatus",
        "type": "object"
      },
...
      "HostKeyCandidate": {
        "additionalProperties": false,
        "description": "等待用户确认或替换的远端 SSH Host Key 候选值。",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "fingerprint_sha256": {
            "pattern": "^SHA256:[A-Za-z0-9+/=_-]+$",
            "title": "Fingerprint Sha256",
            "type": "string"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "type": "string"
          },
          "key_algorithm": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Key Algorithm",
            "type": "string"
          },
          "port": {
            "maximum": 65535,
            "minimum": 1,
            "title": "Port",
            "type": "integer"
          },
          "public_key_openssh_b64": {
            "title": "Public Key Openssh B64",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "host",
          "port",
          "key_algorithm",
          "fingerprint_sha256",
          "public_key_openssh_b64"
        ],
        "title": "HostKeyCandidate",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Expose one bounded machine-readable failure without raw internals.",
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "error_code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "RequestCancelResponse": {
        "additionalProperties": false,
        "description": "Acknowledge an explicit cooperative cancellation request.",
        "properties": {
          "cancellation_requested": {
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "target_request_id": {
            "format": "uuid",
            "title": "Target Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "target_request_id",
          "cancellation_requested"
        ],
        "title": "RequestCancelResponse",
        "type": "object"
      },
      "RuntimeInitializeRequest": {
        "additionalProperties": false,
        "description": "Strict shared input for constructing the one runtime resource graph.",
        "properties": {
          "app_version": {
            "title": "App Version",
            "type": "string"
          },
          "audit_hmac_key_b64": {
            "title": "Audit Hmac Key B64",
            "type": "string"
          },
          "heartbeat_interval_ms": {
            "title": "Heartbeat Interval Ms",
            "type": "integer"
          },
          "heartbeat_timeout_ms": {
            "title": "Heartbeat Timeout Ms",
            "type": "integer"
          },
          "runtime_data_key_b64": {
            "title": "Runtime Data Key B64",
            "type": "string"
          },
          "runtime_db_path": {
            "format": "path",
            "title": "Runtime Db Path",
            "type": "string"
          }
        },
        "required": [
          "app_version",
          "runtime_db_path",
          "runtime_data_key_b64",
          "audit_hmac_key_b64",
          "heartbeat_interval_ms",
          "heartbeat_timeout_ms"
        ],
        "title": "RuntimeInitializeRequest",
        "type": "object"
      },
      "RuntimePhase": {
        "description": "Describe the single runtime owner from live startup through cleanup.",
        "enum": [
          "LIVE_NOT_INITIALIZED",
          "INITIALIZING",
          "READY",
          "DRAINING",
          "CONVERGING",
          "CLOSING",
          "STOPPED",
          "FAILED"
        ],
        "title": "RuntimePhase",
        "type": "string"
      },
      "RuntimeStateResponse": {
        "additionalProperties": false,
        "description": "Return the safe shared runtime lifecycle state.",
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RuntimePhase"
          }
        },
        "required": [
          "request_id",
          "state"
        ],
        "title": "RuntimeStateResponse",
        "type": "object"
      },
...
      "SshConnectRequest": {
        "additionalProperties": false,
        "description": "Freeze target and optional jump profiles with transient credentials.",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "jump": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SshJumpRequest"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "passphrase_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Passphrase B64"
          },
          "password_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Password B64"
          },
          "private_key_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Private Key B64"
          },
          "profile_version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Profile Version",
            "type": "integer"
          }
        },
        "required": [
          "connection_id",
          "profile_version"
        ],
        "title": "SshConnectRequest",
        "type": "object"
      },
      "SshJumpRequest": {
        "additionalProperties": false,
        "description": "Freeze one ProxyJump profile version with its transient credential.",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "passphrase_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Passphrase B64"
          },
          "password_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Password B64"
          },
          "private_key_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Private Key B64"
          },
          "profile_version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Profile Version",
            "type": "integer"
          }
        },
        "required": [
          "connection_id",
          "profile_version"
        ],
        "title": "SshJumpRequest",
        "type": "object"
      },
      "SshSessionListResponse": {
        "additionalProperties": false,
        "description": "Return only safe metadata for active SSH sessions.",
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SshSessionSnapshot"
            },
            "title": "Sessions",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "sessions"
        ],
        "title": "SshSessionListResponse",
        "type": "object"
      },
      "SshSessionSnapshot": {
        "additionalProperties": false,
        "description": "Expose only stable non-secret metadata for one connected SSH session.",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "connection_profile_version": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Connection Profile Version",
            "type": "integer"
          },
          "host_label": {
            "title": "Host Label",
            "type": "string"
          },
          "jump_connection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jump Connection Id"
          },
          "jump_host_key_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jump Host Key Fingerprint"
          },
          "jump_profile_version": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jump Profile Version"
          },
          "ssh_session_id": {
            "format": "uuid",
            "title": "Ssh Session Id",
            "type": "string"
          },
          "target_host_key_fingerprint": {
            "title": "Target Host Key Fingerprint",
            "type": "string"
          }
        },
        "required": [
          "ssh_session_id",
          "connection_id",
          "connection_profile_version",
          "host_label",
          "target_host_key_fingerprint",
          "jump_connection_id",
          "jump_profile_version",
          "jump_host_key_fingerprint"
        ],
        "title": "SshSessionSnapshot",
        "type": "object"
      },
      "SshStatusResponse": {
        "additionalProperties": false,
        "description": "Return one safe SSH lifecycle status.",
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "SshStatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Loopback-only API owned by the Rust-managed packaged child. The Runtime WebSocket is specified separately.",
    "title": "Harness Shell Private Python Runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/requests/{request_id}/cancel": {
      "post": {
        "description": "Explicitly request cooperative cancellation of one active dispatcher call.",
        "operationId": "cancel_request_v1_requests__request_id__cancel_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCancelResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Cancel Request"
      }
    },
    "/v1/runtime/initialize": {
      "post": {
        "description": "Atomically initialize the unique RuntimeResources graph exactly once.",
        "operationId": "initialize_runtime_v1_runtime_initialize_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeInitializeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStateResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Initialize Runtime"
      }
    },
    "/v1/runtime/shutdown": {
      "post": {
        "description": "Converge the ready runtime without bypassing its owner order.",
        "operationId": "shutdown_runtime_v1_runtime_shutdown_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStateResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Shutdown Runtime"
      }
    },
    "/v1/runtime/state": {
      "get": {
        "description": "Return the current safe lifecycle state without resource details.",
        "operationId": "runtime_state_v1_runtime_state_get",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStateResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Runtime State"
      }
    },
...
    "/v1/ssh/sessions": {
      "get": {
        "description": "Return safe metadata snapshots without AsyncSSH transport objects.",
        "operationId": "list_ssh_sessions_v1_ssh_sessions_get",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SshSessionListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "List Ssh Sessions"
      },
      "post": {
        "description": "Use transient credentials to establish one verified SSH session.",
        "operationId": "connect_ssh_session_v1_ssh_sessions_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SshConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SshStatusResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Connect Ssh Session"
      }
    },
    "/v1/ssh/sessions/{ssh_session_id}": {
      "delete": {
        "description": "Close one SSH session and all of its child channels.",
        "operationId": "disconnect_ssh_session_v1_ssh_sessions__ssh_session_id__delete",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "ssh_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ssh Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SshStatusResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Disconnect Ssh Session"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "8765"
        }
      }
    }
  ]
}